Fix nested structured arrays and move unit tests
