+++ numba/cuda/api_util.py
@@ -2,10 +2,6 @@ import numpy as np
 
 
 def prepare_shape_strides_dtype(shape, strides, dtype, order):
-    return _prepare_shape_strides_dtype(shape, strides, dtype, order)
-
-
-def _prepare_shape_strides_dtype(shape, strides, dtype, order):
     dtype = np.dtype(dtype)
     if isinstance(shape, int):
         shape = (shape,)
